Package kann jetzt mit testing.py gestartet werden
