+++ testing.py
@@ -1,0 +1,3 @@
+from knotnpunkt import app
+
+app.app.run()
